+++ src/main/java/org/betonquest/betonquest/quest/event/party/PartyEvent.java
@@ -23,6 +23,11 @@ public class PartyEvent implements Event {
     private final VariableNumber range;
 
     /**
+     * The optional maximum amount of players affected by this party event.
+     */
+    private final VariableNumber amount;
+
+    /**
      * The conditions that must be met by the party members.
      */
     private final ConditionID[] conditions;
@@ -36,11 +41,14 @@ public class PartyEvent implements Event {
      * Creates a new PartyEvent instance.
      *
      * @param range      the range of the party
+     * @param amount     the optional maximum amount of players affected by this party,
+     *                   null or negative values sets no maximum amount
      * @param conditions the conditions that must be met by the party members
      * @param events     the events to fire
      */
-    public PartyEvent(final VariableNumber range, final ConditionID[] conditions, final EventID... events) {
+    public PartyEvent(final VariableNumber range, final VariableNumber amount, final ConditionID[] conditions, final EventID... events) {
         this.range = range;
+        this.amount = amount;
         this.conditions = conditions.clone();
         this.events = events.clone();
     }
@@ -49,10 +57,20 @@ public class PartyEvent implements Event {
     public void execute(final Profile profile) throws QuestRuntimeException {
         final List<OnlineProfile> members = Utils.getParty(profile.getOnlineProfile().get(), range.getDouble(profile),
                 conditions);
+        int toExecute;
+        if (amount != null) {
+            toExecute = amount.getInt(profile);
+        } else {
+            toExecute = -1;
+        }
         for (final OnlineProfile member : members) {
+            if (toExecute == 0) {
+                return;
+            }
             for (final EventID event : events) {
                 BetonQuest.event(member, event);
             }
+            toExecute--;
         }
     }
 }