+++ src/main/java/org/betonquest/betonquest/quest/event/party/PartyEventFactory.java
@@ -25,6 +25,8 @@ public class PartyEventFactory implements EventFactory {
         final VariableNumber range = instruction.getVarNum();
         final ConditionID[] conditions = instruction.getList(instruction::getCondition).toArray(new ConditionID[0]);
         final EventID[] events = instruction.getList(instruction::getEvent).toArray(new EventID[0]);
-        return new OnlineProfileRequiredEvent(new PartyEvent(range, conditions, events), instruction.getPackage());
+        final String amountString = instruction.getOptional("amount");
+        final VariableNumber amount = amountString != null ? instruction.getVarNum(amountString) : null;
+        return new OnlineProfileRequiredEvent(new PartyEvent(range, amount, conditions, events), instruction.getPackage());
     }
 }