+++ CHANGELOG.md
@@ -101,6 +101,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `item` variable - new `name` and `lore` argument and now supports items from other packages
 - `randomnumber` variable added - supports variables
 - `give` event - new `backpack` argument to place items in the backpack (if a valid QuestItem)
+- `party` event - new optional `amount` of maximal affected players
 - Things that are also added in 1.12.X:
     - new line support for `journal_lore` in `messages.yml`
     - FastAsyncWorldEdit compatibility